sap_laba2/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
pandas==2.1.4
python-multipart==0.0.6
jinja2==3.1.2